Refactoring - 코드 수정

+++ SeSAC_Assignment_0131/SeSAC_Assignment_0131/Model/TVDetailsModel.swift
@@ -12,7 +12,7 @@ struct TVDetailsModel: Decodable {
     let name: String
     let numberOfEpisodes: Int
     let overview: String
-    let voteAverage: Double
+    let voteAverage: Float
     
     enum CodingKeys: String, CodingKey {
         case name, overview
